change int types to 8/16 bit

+++ firmware/sim/sim.h
@@ -1,9 +1,10 @@
+#include <stdint.h>
 
 #define LED_WIDTH	4
 #define LED_HEIGHT	5
 
-void setLedXY(int x, int y, int b);
-void setTickInterval(int t);
+void setLedXY(uint8_t x, uint8_t y, uint8_t b);
+void setTickInterval(uint16_t t);
 
 void init();
 void tick();
